Add files via upload
LOOP_CORREG_w_Trans.m is the same as LOOP_CORREGISTRATION.m but allows the user to test various transformations when aligning the fluorescent images (first an x translation and then a y translation) until the user has chosen an alignment they are satisfied with. 

LOOP_CORREG_chicken.m is the same as LOOP_CORREGISTRATION.m but it only uses two images to coregister and not four. I used it when coregistering the chicken images.